--- repositories/src/main/java/RepositoriesConfig.java
@@ -1,2 +1,0 @@
-public class RepositoriesConfig {
-}
