+++ src/main/java/com/ega/datarepositorysevice/repository/ServiceInfoRepository.java
@@ -1,6 +1,7 @@
-package webapp;
+package com.ega.datarepositorysevice.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import webapp.ServiceInfo;
 
 public interface ServiceInfoRepository extends JpaRepository<ServiceInfo, String> {
     ServiceInfo findDistinctFirstByVersion();
