+++ src/test/java/com/ega/datarepositorysevice/DataRepositorySeviceApplicationTests.java
@@ -1,4 +1,4 @@
-package test.java.com.ega.datarepositorysevice;
+package com.ega.datarepositorysevice;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
